+++ anymail/webhooks/postmark.py
@@ -1,8 +1,9 @@
 import json
+import warnings
 
 from django.utils.dateparse import parse_datetime
 
-from ..exceptions import AnymailConfigurationError
+from ..exceptions import AnymailConfigurationError, AnymailWarning
 from ..inbound import AnymailInboundMessage
 from ..signals import (
     AnymailInboundEvent,
@@ -169,13 +170,31 @@ class PostmarkInboundWebhookView(PostmarkBaseWebhookView):
         attachments = [
             AnymailInboundMessage.construct_attachment(
                 content_type=attachment["ContentType"],
-                content=attachment["Content"],
+                content=(
+                    attachment.get("Content")
+                    # WORKAROUND:
+                    # The test webhooks are not like their real webhooks
+                    # This allows the test webhooks to be parsed.
+                    or attachment["Data"]
+                ),
                 base64=True,
                 filename=attachment.get("Name", "") or None,
                 content_id=attachment.get("ContentID", "") or None,
             )
             for attachment in esp_event.get("Attachments", [])
         ]
+
+        # Warning to the user regarding the workaround of above.
+        for attachment in esp_event.get("Attachments", []):
+            if "Data" in attachment:
+                warnings.warn(
+                    "Received a test webhook attachment. "
+                    "It is recommended to test with real inbound events. "
+                    "See https://github.com/anymail/django-anymail/issues/304 "
+                    "for more information.",
+                    AnymailWarning,
+                )
+                break
 
         message = AnymailInboundMessage.construct(
             from_email=self._address(esp_event.get("FromFull")),